file rename: get_player_trip.rb -> get_trip_list.rb
