+++ test/index.rb
@@ -1,6 +1,6 @@
 require_relative "#{__dir__}/../api/get_player_list"
 require_relative "#{__dir__}/../api/calc_play_count"
-require_relative "#{__dir__}/../api/get_player_trip"
+require_relative "#{__dir__}/../api/get_trip_list"
 
 require 'csv'
